# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthUser {
  id: number;
  fullName: string;
  email: string;
  phone: string;
}

export interface SignupInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface CommercialRow {
  description: string;
  timeline: string;
  totalCost: string;
}

export interface DocPropertyRow {
  action: string;
  name: string;
  date: string;
}

export interface VersionHistoryRow {
  version: string;
  dateReleased: string;
  changeNotice: string;
  remark: string;
}

export interface AcceptanceParty {
  name: string;
  designation: string;
  signature: string;
  date: string;
}

export interface ProposalData {
  proposalTitle: string;
  customerName: string;
  documentProperties: DocPropertyRow[];
  versionHistory: VersionHistoryRow[];
  projectSummary: string;
  scopeOfWork: string;
  outOfScope: string;
  commercialRows: CommercialRow[];
  corporateProfile: string;
  orientStrengths: string;
  customerAcceptance: AcceptanceParty;
  orientAcceptance: AcceptanceParty;
}

export interface ProposalInput {
  businessUnit: string;
  data: ProposalData;
}

export interface Proposal {
  id: number;
  userId: number;
  businessUnit: string;
  data: ProposalData;
  createdAt: string;
  updatedAt: string;
}

export interface ProposalSummary {
  id: number;
  businessUnit: string;
  proposalTitle: string;
  customerName: string;
  createdAt: string;
  updatedAt: string;
}