# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Business Proposal Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: proposals
    description: Proposal operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all proposals for current user
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProposalSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a proposal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProposal
      tags: [proposals]
      summary: Update a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [id, fullName, email, phone]

    SignupInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
      required: [fullName, email, phone, password, confirmPassword]

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    CommercialRow:
      type: object
      properties:
        description:
          type: string
        timeline:
          type: string
        totalCost:
          type: string
      required: [description, timeline, totalCost]

    DocPropertyRow:
      type: object
      properties:
        action:
          type: string
        name:
          type: string
        date:
          type: string
      required: [action, name, date]

    VersionHistoryRow:
      type: object
      properties:
        version:
          type: string
        dateReleased:
          type: string
        changeNotice:
          type: string
        remark:
          type: string
      required: [version, dateReleased, changeNotice, remark]

    AcceptanceParty:
      type: object
      properties:
        name:
          type: string
        designation:
          type: string
        signature:
          type: string
        date:
          type: string
      required: [name, designation, signature, date]

    ProposalData:
      type: object
      properties:
        proposalTitle:
          type: string
        customerName:
          type: string
        documentProperties:
          type: array
          items:
            $ref: "#/components/schemas/DocPropertyRow"
        versionHistory:
          type: array
          items:
            $ref: "#/components/schemas/VersionHistoryRow"
        projectSummary:
          type: string
        scopeOfWork:
          type: string
        outOfScope:
          type: string
        commercialRows:
          type: array
          items:
            $ref: "#/components/schemas/CommercialRow"
        corporateProfile:
          type: string
        orientStrengths:
          type: string
        customerAcceptance:
          $ref: "#/components/schemas/AcceptanceParty"
        orientAcceptance:
          $ref: "#/components/schemas/AcceptanceParty"
      required:
        - proposalTitle
        - customerName
        - documentProperties
        - versionHistory
        - projectSummary
        - scopeOfWork
        - outOfScope
        - commercialRows
        - corporateProfile
        - orientStrengths
        - customerAcceptance
        - orientAcceptance

    ProposalInput:
      type: object
      properties:
        businessUnit:
          type: string
        data:
          $ref: "#/components/schemas/ProposalData"
      required: [businessUnit, data]

    Proposal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessUnit:
          type: string
        data:
          $ref: "#/components/schemas/ProposalData"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, businessUnit, data, createdAt, updatedAt]

    ProposalSummary:
      type: object
      properties:
        id:
          type: integer
        businessUnit:
          type: string
        proposalTitle:
          type: string
        customerName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, businessUnit, proposalTitle, customerName, createdAt, updatedAt]