# lib/api-zod/src/generated/types/acceptanceParty.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface AcceptanceParty {
  name: string;
  designation: string;
  signature: string;
  date: string;
}