# lib/api-zod/src/generated/types/docPropertyRow.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface DocPropertyRow {
  action: string;
  name: string;
  date: string;
}