# lib/api-zod/src/generated/types/proposal.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalData } from './proposalData';

export interface Proposal {
  id: number;
  userId: number;
  businessUnit: string;
  data: ProposalData;
  createdAt: string;
  updatedAt: string;
}