# lib/api-zod/src/generated/types/proposalData.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { AcceptanceParty } from './acceptanceParty';
import type { CommercialRow } from './commercialRow';
import type { DocPropertyRow } from './docPropertyRow';
import type { VersionHistoryRow } from './versionHistoryRow';

export interface ProposalData {
  proposalTitle: string;
  customerName: string;
  documentProperties: DocPropertyRow[];
  versionHistory: VersionHistoryRow[];
  projectSummary: string;
  scopeOfWork: string;
  outOfScope: string;
  commercialRows: CommercialRow[];
  corporateProfile: string;
  orientStrengths: string;
  customerAcceptance: AcceptanceParty;
  orientAcceptance: AcceptanceParty;
}