# lib/api-zod/src/generated/types/proposalSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ProposalSummary {
  id: number;
  businessUnit: string;
  proposalTitle: string;
  customerName: string;
  createdAt: string;
  updatedAt: string;
}