# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Business Proposal Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}